backend/requirements.txt
fastapi
uvicorn
chromadb
trafilatura
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
google-genai
python-dotenv